% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_identity.R
\name{on_connect_cloud}
\alias{on_connect_cloud}
\title{Is this content running on Posit Connect Cloud?}
\usage{
on_connect_cloud()
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Is this content running on Posit Connect Cloud?
}
