% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section_tracking.R
\name{section_tracking_js}
\alias{section_tracking_js}
\title{Get the section tracking JavaScript tag}
\usage{
section_tracking_js()
}
\value{
An \code{\link[htmltools:HTML]{htmltools::HTML()}} string containing the \verb{<script>} block.
}
\description{
Returns an HTML \verb{<script>} tag that listens for URL hash changes (learnr
encodes section navigation in the hash) and fires a Shiny input event.
Include this in the tutorial Rmd before the first section heading.
}
